V1.1.3
More improvemnts on FFI

New FFI animation mode:
"BMS" - FFI is rendered as the BMS one. animated Hundreds, the rest are jumping
"RealFFI" - FFI is rendered as the Real thing, no 0 on higher numbers, rollover on all three top digits. (lower FPS)

both modes include an optional Bezel drawn on top of the screen.
font is set to be at the correct size with the Bezel off. Bezel on reduces the "window" size compared to the real FFI.

Default FFI setting are "RealFFI" with "Bezel".
Note that turning any of these options off will increase FPS and animation smoothness (Bezel being the more expensive of the two)

New font is included, designed for maximum perfromance with minimum memory footprint.

Thank you to "AiRdAncE" and "Moon" from Viperpits and "Topper Harley" of ffw05.fr for helping me improve the FFI.