config.h refracting
All important configuration options are now located on "config.h" file - no need to fuss around in the code
Arduino Micro is now the default controller. Arduino Uno still works, but Bezel and RealFFI are autocratically disabled on it for performance reasons.

Hopefully this will be the last of the SPI based light system releases. so please be aware when designing boards. - current AOA and Caution panel will still be available via SPI as they are now.